refactor functions to dedicated files
